@@ -1,13 +1,13 @@
-import logging
+"""
+tmtoolkit – Text Mining and Topic Modeling Toolkit for Python
 
-from . import topicmod, bow
+Markus Konrad <markus.konrad@wzb.eu>
+"""
+
+from . import topicmod, bow, preprocess
 
 
 __title__ = 'tmtoolkit'
-__version__ = '0.8.0'
+__version__ = '0.9.0-rc1'
 __author__ = 'Markus Konrad'
 __license__ = 'Apache License 2.0'
-
-
-logger = logging.getLogger(__title__)
-logger.addHandler(logging.NullHandler())
